+++ phunk/__init__.py
@@ -3,5 +3,6 @@ from .core import PhaseCurve
 from . import miriade
 from . import models
 from . import plotting
+from .logging import set_log_level
 
 VERSION = "0.1.1"